% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.ts.R
\docType{data}
\name{A}
\alias{A}
\title{A Santa Fe time series}
\description{
chaotic time series from Santa Fe time series competition
}
\details{
Dataset A
}
\examples{
data(A)
plot(A[,1],type="l")
}
\keyword{data}

