% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteria.R
\name{AUC}
\alias{AUC}
\title{AUC}
\usage{
AUC(y, yhat)
}
\arguments{
\item{y:}{real value}

\item{yhat:}{predicted probability}
}
\value{
AUC
}
\description{
AUC
}
\details{
AUC

AUC
}
\examples{
## random prediction
AUC(round(runif(100)),rnorm(100))

}
\references{
Handbook \emph{Statistical foundations of machine learning} available in \url{http://www.ulb.ac.be/di/map/gbonte/mod_stoch/syl.pdf}
}
\author{
Gianluca Bontempi  \email{gbonte@ulb.ac.be}
}
