% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlearn.R
\name{bagging.pred}
\alias{bagging.pred}
\title{Wrapper on lazy learning algoritmhs for regression and classification}
\usage{
bagging.pred(
  algo,
  X,
  Y,
  X.ts,
  B = 10,
  maxN = Inf,
  classi = TRUE,
  balanced = FALSE,
  bagvar = 0,
  ...
)
}
\arguments{
\item{X:}{training input}

\item{Y:}{training output}

\item{X.ts:}{test input}

\item{classi:}{TRUE for classification, FALSE for regression}

\item{return.more:}{TRUE for returning additional \pkg{lazy} parameter}
}
\value{
if \code{classi=FALSE} predicted test output; if \code{classi=TRUE} a list with
\itemize{
\item{\code{pred}:}  predicted class
\item{ \code{prob}:} posteriori probability
}
}
\description{
Bagging wrapper on algoritmhs
}
\references{
\url{mlg.ulb.ac.be}
}
\author{
Gianluca Bontempi  \email{gbonte@ulb.ac.be}
}
