% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biasvar.R
\name{BiasVar}
\alias{BiasVar}
\title{Demo of bias/variance tradeoff}
\usage{
BiasVar(N = 25, n = 20, R = 200, seed = 1)
}
\arguments{
\item{N:}{number of observations}

\item{n:}{max degree of polynomial fitting}

\item{R:}{number of repetitions for bias and variance estimation}

\item{seed:}{seed random generator}
}
\value{
: list with
\itemize{
\item{Remp}: training MSE
\item{B}: estimated squared bias
\item{V}: estimated sqaured variance
}
}
\description{
Demo of bias/variance tradeoff
}
\examples{

BiasVar()



}
\references{
\url{mlg.ulb.ac.be}
}
\author{
Gianluca Bontempi  \email{gbonte@ulb.ac.be}
}
