% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.ts.R
\name{constloo}
\alias{constloo}
\title{Embedding of multivariate time series}
\usage{
constloo(x, w = rep(1, length(x)))
}
\arguments{
\item{x:}{observation vector}

\item{w:}{weight vector}
}
\value{
: leave-one-out Mean Squared Error
}
\description{
Leave-one-out Mean Squared Error of a weighted mean estimator
}
\author{
Gianluca Bontempi  \email{gbonte@ulb.ac.be}
}
\references{
Handbook \emph{Statistical foundations of machine learning} available in \url{http://www.ulb.ac.be/di/map/gbonte/mod_stoch/syl.pdf}
}

