% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{cor2I2}
\alias{cor2I2}
\title{correlation to mutual information}
\usage{
cor2I2(rho)
}
\arguments{
\item{rho:}{Pearson correlation}
}
\value{
mutual information
}
\description{
correlation to mutual information under hypothesis of normality
}
\details{
cor2I2

correlation to mutual information under hypothesis of normality
}
\examples{
cor2I2(-0.9)

}
\author{
Gianluca Bontempi  \email{gbonte@ulb.ac.be}
}
\references{
Handbook \emph{Statistical foundations of machine learning} available in \url{http://www.ulb.ac.be/di/map/gbonte/mod_stoch/syl.pdf}
}

