% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.ts.R
\name{dist2}
\alias{dist2}
\title{Returns matrix of distances between two matrices with same number of columns}
\usage{
dist2(X1, X2)
}
\arguments{
\item{X1:}{matrix [N1,n]}

\item{X2:}{matrix [N2,n]}
}
\value{
: matrix of euclidean distances [N1,N2]
}
\description{
Returns matrix of distances between two matrices with same number of columns
}
\references{
\url{mlg.ulb.ac.be}
}
\author{
Gianluca Bontempi  \email{gbonte@ulb.ac.be}
}
