% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nar.R
\name{genar}
\alias{genar}
\title{NAR univariate time series generation}
\usage{
genar(N, s = 0.1, number = 1)
}
\arguments{
\item{number}{: integer number (between 1 and 13) denoting the NAR generating model}

\item{N:}{number of observations}

\item{s:}{standard deviation of additive noise}
}
\value{
vector containing time series observations
}
\description{
Generte a univariate time series using a NAR mode
}
\examples{
plot(genar(100),type="l")


}
\references{
\url{https://tinyurl.com/sfmlh}
}
\author{
Gianluca Bontempi  \email{gbonte@ulb.ac.be}
}
