% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featuresel.R
\name{gsloo}
\alias{gsloo}
\title{gsloo}
\usage{
gsloo(X, Y, nmax = ncol(X), trace = FALSE, automatic = F, class = F, shave = F)
}
\arguments{
\item{X:}{input dataset}

\item{Y:}{output dataset}

\item{nmax:}{number of top returned features}
}
\value{
Indices of \code{nmax} top ranked features
}
\description{
Ranking filter based on Gram Schmidt orthogonalization
}
\details{
ranking by GS orthogonalization

Ranking filter based on Gram Schmidt orthogonalization
}
\examples{
N<-100
n<-5
neff<-3
R<-regrDataset(N,n,neff,0.1,seed=0)
X<-R$X
Y<-R$Y
real.features<-R$feat
ranked.features<-gsloo(X,Y,nmax=3)
}
\references{
Handbook \emph{Statistical foundations of machine learning} available in \url{https://tinyurl.com/sfmlh}
}
\author{
Gianluca Bontempi  \email{Gianluca.Bontempi@ulb.be}
}
