% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{hyperg}
\alias{hyperg}
\title{hypergeometric pvalue}
\usage{
hyperg(m, n, k, x)
}
\arguments{
\item{m:}{number of marked elements}

\item{n:}{number of unmarked elements}

\item{k:}{selection size}

\item{x:}{number of marked elements in our selection}
}
\value{
p: pvalue. Probability of selecting a number of marked elements larger equal than x if the selection if random
}
\description{
Returns pvalue of hypergeometric
}
\details{
hypergeometric pvalue

Returns pvalue of hypergeometric
}
\examples{
hyperg(10,5,2,1)

}
\author{
Gianluca Bontempi  \email{gbonte@ulb.ac.be}
}
\references{
Handbook \emph{Statistical foundations of machine learning} available in \url{http://www.ulb.ac.be/di/map/gbonte/mod_stoch/syl.pdf}
}

