% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{impute}
\alias{impute}
\alias{imputeDataset}
\title{impute}
\usage{
imputeDataset(X)
}
\arguments{
\item{X:}{dataset}
}
\value{
A list with:
\itemize{
\item{impX:} imputed dataset
\item{remcol:} indices of remaining columns
}
}
\description{
Impute data set by removing constant columns and replacing NA samples with the mean
}
\details{
imputeDataset

Impute data set by removing constant columns and replacing NA samples with the mean
}
\author{
Gianluca Bontempi  \email{gbonte@ulb.ac.be}
}
\references{
Handbook \emph{Statistical foundations of machine learning} available in \url{http://www.ulb.ac.be/di/map/gbonte/mod_stoch/syl.pdf}
}

