% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{is.element.list}
\alias{is.element.list}
\title{is.element.list}
\usage{
is.element.list(e, l)
}
\arguments{
\item{e:}{element}

\item{l:}{list}
}
\description{
Return TRUE if element belongs to list
}
\details{
is.element.list

Return TRUE if element belongs to list
}
\examples{
L<-list("aa","b","a1","a","E")
is.element.list("a",L)
}
\author{
Gianluca Bontempi  \email{gbonte@ulb.ac.be}
}
\references{
Handbook \emph{Statistical foundations of machine learning} available in \url{http://www.ulb.ac.be/di/map/gbonte/mod_stoch/syl.pdf}
}

