% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlearn.R
\name{lazy.pred}
\alias{lazy.pred}
\title{Wrapper on lazy learning algoritmhs for regression and classification}
\usage{
lazy.pred(
  X,
  Y,
  X.ts,
  class = FALSE,
  return.more = FALSE,
  conPar = NULL,
  linPar = NULL,
  cmbPar = 10,
  scaleX = TRUE
)
}
\arguments{
\item{X:}{training input}

\item{Y:}{training output}

\item{X.ts:}{test input}

\item{conPar:}{constant modeling parameters}

\item{linPar:}{linear modeling parameters}

\item{cmbPar:}{combination parameters}

\item{classi:}{TRUE for classification, FALSE for regression}

\item{return.more:}{TRUE for returning additional \pkg{lazy} parameter}
}
\value{
if \code{classi=FALSE} predicted test output; if \code{classi=TRUE} a list with
\itemize{
\item{\code{pred}:}  predicted class
\item{ \code{prob}:} posteriori probability
}
}
\description{
Wrapper on lazy learning algoritmhs
}
\references{
\url{mlg.ulb.ac.be}
}
\author{
Gianluca Bontempi  \email{gbonte@ulb.ac.be}
}
