% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstep.R
\name{lin.pls}
\alias{lin.pls}
\title{lin.pls}
\description{
Multioutput lin PLS
}
\details{
lin.pls

Multioutput PLS for multi-step-ahed prediction. It performs a  partial least-squares
}
\references{
\emph{Bontempi G. Ben Taieb S. Conditionally dependent strategies for multiple-step-ahead prediction in local learning, International Journal of Forecasting Volume 27, Issue 3, July–September 2011, Pages 689–699}
}
\author{
Gianluca Bontempi  \email{gbonte@ulb.ac.be}
}
