% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featuresel.R
\name{linearFsel}
\alias{linearFsel}
\title{linearFsel}
\usage{
linearFsel(X, Y, nmax = 5, nmax2 = NCOL(X), loo = FALSE, back = FALSE)
}
\arguments{
\item{X:}{input dataset}

\item{Y:}{output dataset}

\item{nmax:}{number of top returned features}

\item{nmax2:}{number of forward selection steps}

\item{back:}{if TRUE, backward reordering based on linear regression}

\item{loo:}{if TRUE, assessment based on leave-one-out MSE}
}
\value{
Indices of \code{nmax} top ranked features
}
\description{
Wrapper feature selection based on forward selection and linear regression
}
\details{
linearFsel

Wrapper feature selection based on forward selection and linear regression
}
\examples{
N<-100
n<-5
neff<-3
R<-regrDataset(N,n,neff,0.1,seed=0)
X<-R$X
Y<-R$Y
real.features<-R$feat
ranked.features<-linearFsel(X,Y,nmax=3)
}
\references{
Handbook \emph{Statistical foundations of machine learning} available in \url{https://tinyurl.com/sfmlh}
}
\author{
Gianluca Bontempi  \email{Gianluca.Bontempi@ulb.be}
}
