% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.ts.R
\name{MakeEmbedded}
\alias{MakeEmbedded}
\title{Embedding of multivariate time series}
\usage{
MakeEmbedded(ts, n, delay, hor = 1, w = 1)
}
\arguments{
\item{n}{[no.var]: vector of embedding orders}

\item{delay}{[no.var]: vector of delays}

\item{hor}{[no.var]: vector of predicted horizons (hor=1 boils down to one-step-ahed prediction)}

\item{ts:}{multivariate time series [no. observations,no. variates]}

\item{w:}{index of variables appearing in $out}
}
\value{
list with
\itemize{
\item{inp}: embedded inputs of all variables (n[i] columns for series ts[i])
\item{out}: outputs of variates whose index is in w}
}
\description{
Embed a multivariate time series in input output form
}
\examples{
TS<-array(1:25,c(5,5))
## embeds a 5 variate time series with embedding orders equal to 2 for all variables for one-step ahead prediction
MakeEmbedded(TS,n=rep(2,5),delay=rep(0,5),hor=rep(1,5),w=1:5)


}
\references{
\url{mlg.ulb.ac.be}
}
\author{
Gianluca Bontempi  \email{gbonte@ulb.ac.be}
}
