% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.ts.R
\name{MASE}
\alias{MASE}
\title{Mean Absolute scaled error}
\usage{
MASE(y, yhat)
}
\description{
Mean Absolute scaled error
}
\author{
Gianluca Bontempi  \email{gbonte@ulb.ac.be}
}
\references{
\url{https://en.wikipedia.org/wiki/Mean_absolute_scaled_error}
}

