% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featuresel.R
\name{mimr (minimum interaction maximum relevance)}
\alias{mimr}
\alias{mimr (minimum interaction maximum relevance)}
\title{mimr}
\usage{
mimr(X, Y, nmax = 5, first = NULL, init = FALSE, lambda = 0.5,
  fast.inter = 0, back = FALSE, spouse.removal = TRUE, caus = 1)
}
\arguments{
\item{X:}{input dataset}

\item{Y:}{output dataset}

\item{nmax:}{number of top returned features}

\item{back:}{if TRUE, backward reordering based on linear regression}

\item{caus:}{if caus =1 it searches for causes otherwise if caus=-1 it searches for effects}
}
\value{
Indices of \code{nmax} top ranked features
}
\description{
mimr filter based on mutual information
}
\details{
mimr

mimr (minimum interaction maximum relevance) filter based on mutual information
}
\examples{
N<-100
n<-5
neff<-3
R<-regrDataset(N,n,neff,0.1,seed=0)
X<-R$X
Y<-R$Y
real.features<-R$feat
ranked.features<-mimr(X,Y,nmax=3)
}
\author{
Gianluca Bontempi  \email{gbonte@ulb.ac.be}
}
\references{
Handbook \emph{Statistical foundations of machine learning} available in \url{http://www.ulb.ac.be/di/map/gbonte/mod_stoch/syl.pdf}
}

