% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featuresel.R
\name{mrmr}
\alias{mrmr}
\title{mrmr}
\usage{
mrmr(X, Y, nmax = 5, back = FALSE)
}
\arguments{
\item{X:}{input dataset}

\item{Y:}{output dataset}

\item{nmax:}{number of top returned features}

\item{back:}{if TRUE, backward reordering based on linear regression}
}
\value{
Indices of \code{nmax} top ranked features
}
\description{
mRMR filter based on mutual information
}
\details{
mrmr

mRMR (minimum redundancy maximum relevance) filter based on mutual information
}
\examples{
N<-100
n<-5
neff<-3
R<-regrDataset(N,n,neff,0.1,seed=0)
X<-R$X
Y<-R$Y
real.features<-R$feat
ranked.features<-mrmr(X,Y,nmax=3)
}
\author{
Gianluca Bontempi  \email{gbonte@ulb.ac.be}
}
\references{
Handbook \emph{Statistical foundations of machine learning} available in \url{http://www.ulb.ac.be/di/map/gbonte/mod_stoch/syl.pdf}
}

