% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.ts.R
\docType{data}
\name{NN5}
\alias{NN5}
\title{NN5 time series}
\description{
time series from NN5 time series competition
}
\details{
Dataset NN5
}
\examples{
data(NN5)
plot(NN5[,1],type="l")
}
\references{
\url{http://www.neural-forecasting-competition.com/NN5/}

\url{https://www.researchgate.net/publication/222828050_Conditionally_dependent_strategies_for_multiple-step-ahead_prediction_in_local_learning}
}
\keyword{data}
