% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{pca.svd}
\alias{pca.svd}
\title{PCA compression by SVD decomposition}
\usage{
pca.svd(X, n.comp = 2)
}
\arguments{
\item{X:}{input data}

\item{n.comp:}{number of returned principal components}
}
\value{
list
\itemize{
\item{data}: transformed data
\item{vect}: principal eigenvectors
\item{val}: principal eigenvalues
\item{ord}:
\item{mu}: mean of original data
}
}
\description{
PCA compression by SVD decomposition
}
\details{
PCA by SVD decomposition

PCA compression by SVD decomposition
}
\examples{
N<-100
n<-5
neff<-3
R<-regrDataset(N,n,neff,0.1)
X<-R$X
Z<-pca.svd(X,3)$data

}
\author{
Gianluca Bontempi  \email{gbonte@ulb.ac.be}
}
\references{
Handbook \emph{Statistical foundations of machine learning} available in \url{http://www.ulb.ac.be/di/map/gbonte/mod_stoch/syl.pdf}
}

