% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{pcor1}
\alias{pcor1}
\title{partial correlation}
\usage{
pcor1(x, y, z)
}
\arguments{
\item{x:}{vector}

\item{y:}{vector}

\item{z:}{vector}
}
\value{
partial correlation
}
\description{
partial correlation cor(x,y|z)
}
\details{
pcor1

partial correlation cor(x,y|z)
}
\examples{
pcor1(rnorm(100),rnorm(100),rnorm(100),)

}
\references{
Handbook \emph{Statistical foundations of machine learning} available in \url{http://www.ulb.ac.be/di/map/gbonte/mod_stoch/syl.pdf}
}
\author{
Gianluca Bontempi  \email{gbonte@ulb.ac.be}
}
