% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStu.R
\name{plotStu}
\alias{plotStu}
\title{Plot the Student density and cumulative distribution}
\usage{
plotStu(N = 10)
}
\description{
Plot Student distribution
}
\author{
Gianluca Bontempi  \email{gbonte@ulb.ac.be}
}
\references{
\url{mlg.ulb.ac.be}
}

