% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{preproc2}
\alias{preproc2}
\title{preproc2}
\usage{
preproc2(
  X1,
  X2,
  remove.nan = TRUE,
  remove.const = TRUE,
  to.scale = TRUE,
  verbose = T,
  cluster = 0
)
}
\arguments{
\item{X1:}{training dataset}

\item{X2:}{test dataset}

\item{remove.nan:}{TRUE/FALSE. It removes NaN}

\item{remove.const:}{TRUE/FALSE. It removes constant columns.}

\item{to.scale:}{TRUE/FALSE. It normalizes columns}

\item{cluster:}{Number of feature clusters. If =0 it does nothing.}
}
\value{
A list with:
\itemize{
\item{X1:} preprocessed training dataset
\item{X2:} preprocessed training dataset
\item{remcol:} indices of remaining columns
\item{groups:} indices of clusters components
}
}
\description{
Preprocesses a training and a test dataset
}
\details{
preproc2

It performs hierarchical clustering to reduce the number of features
}
\references{
Handbook \emph{Statistical foundations of machine learning}
}
\author{
Gianluca Bontempi  \email{gbonte@ulb.ac.be}
}
