% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featuresel.R
\name{rankregr}
\alias{rankregr}
\title{rankregr}
\usage{
rankregr(X, Y, lambda = 0.01, nmax = 5)
}
\arguments{
\item{X:}{input dataset}

\item{Y:}{output dataset}

\item{nmax:}{number of top returned features}
}
\value{
Indices of \code{nmax} top ranked features
}
\description{
Feature ranking based on linear least squares regression coefficients
}
\details{
rankregr

Feature ranking based on linear least squares regression coefficients
}
\examples{
N<-100
n<-5
neff<-3
R<-regrDataset(N,n,neff,0.1,seed=0)
X<-R$X
Y<-R$Y
real.features<-R$feat
ranked.features<-rankregr(X,Y,nmax=3)
}
\references{
Handbook \emph{Statistical foundations of machine learning} available in \url{https://tinyurl.com/sfmlh}
}
\author{
Gianluca Bontempi  \email{Gianluca.Bontempi@ulb.be}
}
