% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featuresel.R
\name{rankrho}
\alias{rankrho}
\title{rankrho}
\usage{
rankrho(X, Y, nmax = 5, regr = FALSE, first = NULL)
}
\description{
value
}
\details{
Rankrho function

aa
}
\author{
Gianluca Bontempi  \email{gbonte@ulb.ac.be}
}
\references{
\url{mlg.ulb.ac.be}
}
\keyword{internal}

