% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependency.R
\name{regrDataset}
\alias{regrDataset}
\title{Generator of nonlinear regression dataset}
\usage{
regrDataset(N, n, neff, sdn, seed = 0)
}
\arguments{
\item{N:}{number of observations}

\item{n:}{number of observed variables}

\item{neff:}{number of effective features}

\item{sdn:}{standard deviation of Gaussian noise}

\item{seed:}{seed random generator}
}
\examples{

N<-100
n<-5
neff<-3
R<-regrDataset(N,n,neff,0.1)
X<-R$X
Y<-R$Y
}
\references{
\url{mlg.ulb.ac.be}
}
\author{
Gianluca Bontempi  \email{gbonte@ulb.ac.be}
}
