% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin.R
\name{regrlin}
\alias{regrlin}
\title{Linear regression with PRESS leave one out estimate}
\usage{
regrlin(X, Y, X.ts = NULL, lambda = 0.001)
}
\arguments{
\item{X:}{training input}

\item{Y:}{training output}

\item{X.ts:}{test input}

\item{lambda:}{regularization parameter}
}
\value{
a list with fields:
\itemize{
\item{\code{e}}: training error,
\item{ \code{beta.hat}:} coeffcients,
\item{ \code{MSE.emp}:} training MSE,
\item{ \code{sdse.emp}:} standard deviation of squared error,
\item{ \code{var.hat}:} estimated noise variance;
\item{ \code{MSE.loo}:} PRESS MSE,
\item{ \code{sdse.loo}:} standard deviation of squared leave-one-out error error,
\item{ \code{Y.hat}:} predicted training output,
\item{ \code{Y.hat.ts}:} predicted test output,
\item{ \code{e.loo}:} leave-one-out error
}
}
\description{
Linear regression
}
\examples{
N<-100
n<-10
X<-array(rnorm(N*n),c(N,n))
Y<-sin(X[,1]*X[,2])
R<-regrlin(X,Y,X)
R2<-regrlin.d(X,Y,X,lambda=1e-5)


}
\references{
\url{mlg.ulb.ac.be}
}
\author{
Gianluca Bontempi  \email{gbonte@ulb.ac.be}
}
