% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.ts.R
\name{remNA}
\alias{remNA}
\title{Remove NA from a time series by interpolation}
\usage{
remNA(TS)
}
\arguments{
\item{TS:}{univariate time series}
}
\description{
Remove NA from a time series by interpolation
}
\references{
\url{mlg.ulb.ac.be}
}
\author{
Gianluca Bontempi  \email{gbonte@ulb.ac.be}
}
