% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featuresel.R
\name{rfrank}
\alias{rfrank}
\title{rfrank}
\usage{
rfrank(X, Y, nmax = 5, type = 1)
}
\arguments{
\item{X:}{input dataset}

\item{Y:}{output dataset}

\item{nmax:}{number of top returned features}

\item{type:}{importance type: classification (1:MeanDecreaseAccuracy, 2: MeanDecreaseGini), regression (1:%IncMSE, 2:IncNodePurity)}
}
\value{
Indices of \code{nmax} top ranked features
}
\description{
rfrank filter based on importance of Random Forest
}
\details{
rfrank

rfrank based on importance of Random Forest
}
\examples{
N<-100
n<-5
neff<-3
R<-regrDataset(N,n,neff,0.1,seed=0)
X<-R$X
Y<-R$Y
real.features<-R$feat
ranked.features<-rfrank(X,Y,nmax=3)


}
\references{
Handbook \emph{Statistical foundations of machine learning} available in \url{https://tinyurl.com/sfmlh}
}
\author{
Gianluca Bontempi  \email{Gianluca.Bontempi@ulb.be}
}
