% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featuresel.R
\name{ridgeFsel}
\alias{ridgeFsel}
\title{ridgeFsel}
\usage{
ridgeFsel(X, Y, nmax = 5, alpha = 0)
}
\arguments{
\item{X:}{input dataset}

\item{Y:}{output dataset}

\item{nmax:}{number of top returned features}
}
\description{
Ridge based feature selection based on glmnet
}
\details{
ridgeFsel

Ridge based feature selection based on glmnet
}
\references{
Handbook \emph{Statistical foundations of machine learning} available in \url{https://tinyurl.com/sfmlh}
}
\author{
Gianluca Bontempi  \email{Gianluca.Bontempi@ulb.be}
}
