% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.ts.R
\name{SMAPE}
\alias{SMAPE}
\title{Symmetric mean absolute percentage error}
\usage{
SMAPE(ts1, ts2, Cent = 0, Sc = 1)
}
\description{
Symmetric mean absolute percentage error
}
\references{
\url{https://en.wikipedia.org/wiki/Symmetric_mean_absolute_percentage_error}
}
\author{
Gianluca Bontempi  \email{gbonte@ulb.ac.be}
}
