% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{StatPredictors1}
\alias{StatPredictors1}
\title{Statistical forecasting predictors}
\usage{
StatPredictors1(input, fh, index = 0, verbose = F)
}
\arguments{
\item{input:}{time series}

\item{fh:}{horizon}
}
\value{
: forecasting vector of size fh
}
\description{
Statistical forecasting predictors
}
\references{
M4 competition
}
\author{
Gianluca Bontempi  \email{gbonte@ulb.ac.be}
}
