% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstep.R
\name{timefit}
\alias{timefit}
\title{timefit}
\usage{
timefit(TS.tr, n, C, H)
}
\arguments{
\item{TS.tr:}{training time series}

\item{C:}{number of fitted points}

\item{H:}{horizon}
}
\value{
vector of H predictions
}
\description{
time fitting of time series
}
\details{
timefit

time fitting
}
\author{
Gianluca Bontempi  \email{Gianluca.Bontempi@ulb.be}
}
