% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{var2H}
\alias{var2H}
\title{variance to entropy}
\usage{
var2H(v)
}
\arguments{
\item{v:}{variance}
}
\value{
entropy
}
\description{
variance to entropy under hypothesis of normality
}
\details{
var2H

variance to entropy under hypothesis of normality
}
\examples{
var2H(-0.9)

}
\author{
Gianluca Bontempi  \email{gbonte@ulb.ac.be}
}
\references{
Handbook \emph{Statistical foundations of machine learning} available in \url{http://www.ulb.ac.be/di/map/gbonte/mod_stoch/syl.pdf}
}

