% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{which.element.list}
\alias{which.element.list}
\title{which.element.list}
\usage{
which.element.list(e, l)
}
\arguments{
\item{e:}{element}

\item{l:}{list}
}
\description{
Return index list equal to a a given element
}
\details{
which.element.list

Return index of list whose content is equal to a a given element
}
\examples{
L<-list("aa","b","a1","a","E")
which.element.list("a",L)

}
\references{
Handbook \emph{Statistical foundations of machine learning} available in \url{http://www.ulb.ac.be/di/map/gbonte/mod_stoch/syl.pdf}
}
\author{
Gianluca Bontempi  \email{gbonte@ulb.ac.be}
}
