% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{which.subelement.list}
\alias{which.subelement.list}
\title{which.subelement.list}
\usage{
which.subelement.list(e, l)
}
\arguments{
\item{e:}{element}

\item{l:}{list}
}
\description{
Returns indices of list terms xhich contain a given element
}
\details{
which.subelement.list

Returns indices of list terms xhich contain a given element
}
\examples{
L<-list("aa","b","a1","a","E")
L[[6]]<-c("a","z")
which.subelement.list("a",L)

}
\references{
Handbook \emph{Statistical foundations of machine learning} available in \url{http://www.ulb.ac.be/di/map/gbonte/mod_stoch/syl.pdf}
}
\author{
Gianluca Bontempi  \email{gbonte@ulb.ac.be}
}
