% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.ts.R
\docType{data}
\name{A}
\alias{A}
\title{A Santa Fe time series}
\description{
chaotic time series from Santa Fe time series competition
}
\details{
Dataset A
}
\examples{
data(A)
plot(A[,1],type="l")
}
\references{
\url{http://www-psych.stanford.edu/~andreas/Time-Series/SantaFe.html}

\url{http://www.amazon.com/Time-Series-Prediction-Forecasting-Understanding/dp/0201626020}
}
\keyword{data}
