% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstep.R
\name{multiplestepAhead}
\alias{multiplestepAhead}
\title{multiplestepAhead}
\arguments{
\item{TS:}{time series}

\item{n:}{embedding order}

\item{H:}{horizon}

\item{Kmin:}{min number of neighbours}

\item{dist:}{type of distance: \code{euclidean, cosine}}

\item{F:}{forgetting factor}

\item{C:}{integer parameter which sets the maximum number of neighbours (C*k)}

\item{detrend:}{real parameter (in [0,1]) which fixes the window used for detrending the series. If equal to -1, no detrending is carried out}

\item{smooth:}{if TRUE, the preidction is obtained by averaging multiple windows with different starting points}

\item{method:}{\itemize{
\item{arima}: prediction based on the \pkg{forecast} package
\item{direct}: direct prediction based on \link{KNN.multioutput} function
\item{iter}: recursive prediction based on \link{KNN.multioutput} function
\item{lazydirect}: direct prediction based on \link{lazy.pred} function
\item{lazyiter}: recursive prediction based on \link{lazy.pred} function
\item{rfdirect}: direct prediction based on \link{rf.pred} function
\item{rfiter}: recursive prediction based on \link{rf.pred} function
\item{mimo}: MIMO prediction based on \link{KNN.multioutput} function
\item{mimo.comb}: MIMO prediction based on \link{KNN.multioutput} function which combines a set of predictors based on different horizons and different starting points
\item{mimo.acf}: MIMO prediction based on \link{KNN.acf} function which combines a set of predictors based on different horizons and different starting points
\item{mimo.acf.lin}: MIMO prediction based on \link{KNN.acf.lin} function which combines a set of predictors based on different horizons and different starting points
\item{mimo.pls}: MIMO prediction based on \link{KNN.pls} function which combines a set of predictors based on different horizons and different starting points
\item{mimo.lin.pls}: MIMO prediction based on Partial Least Squares which combines a set of predictors based on different horizons and different starting points
}}
}
\value{
H step ahead predictions
}
\description{
multiplestepAhead
}
\details{
multiplestepAhead

Wrapper over a set of methods for multiple step ahead time series prediction
}
\examples{
## Multi-step ahead time series forecasting

t=seq(0,400,by=0.1)
N<-length(t)
H<-500 ## horizon prediction
TS<-sin(t)+rnorm(N,sd=0.1)
TS.tr=TS[1:(N-H)]
N.tr<-length(TS.tr)
TS.ts<-TS[(N-H+1):N]
TS.tr=array(TS.tr,c(length(TS.tr),1))
Y.cont=multiplestepAhead(TS.tr,n=3, H=H,method="mimo")

plot(t[(N-H+1):N],TS.ts)
lines(t[(N-H+1):N],Y.cont)




## Multi-step ahead time series forecasting Santa Fe  chaotic time series A
rm(list=ls()
data(A)
TS=A
N<-1000
H<-200
TS.tr=TS[1:N,1]
TS.ts<-TS[(N+1):(N+H),1]
TS.tr=array(TS.tr,c(length(TS.tr),1))
Y.mimo=multiplestepAhead(TS.tr,n=12, H=H,method="mimo")
Y.mimo.comb=multiplestepAhead(TS.tr,n=12, H=H,method="mimo.comb")
plot(t[(N-H+1):N],TS.ts)
lines(t[(N-H+1):N],Y.mimo,col="red")
lines(t[(N-H+1):N],Y.mimo.comb,col="green")



}
\references{
\emph{Bontempi G. Ben Taieb S. Conditionally dependent strategies for multiple-step-ahead prediction in local learning, International Journal of Forecasting Volume 27, Issue 3, July–September 2011, Pages 689–699}
}
\author{
Gianluca Bontempi  \email{gbonte@ulb.ac.be}
}
