% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstep.R
\name{multiplestepAhead}
\alias{multiplestepAhead}
\title{multiplestepAhead}
\usage{
multiplestepAhead(TS, n, H, D = 0, method = "direct", Kmin = 3, C = 2,
  FF = 0)
}
\arguments{
\item{TS:}{time series}

\item{n:}{embedding order}

\item{H:}{horizon}

\item{Kmin:}{min number of neighbours}

\item{dist:}{type of distance: \code{euclidean, cosine}}

\item{F:}{forgetting factor}

\item{C:}{integer parameter which sets the maximum number of neighbours (Ck)}

\item{method:}{\itemize{
\item{arima}: prediction based on the \pkg{forecast} package
\item{direct}: direct prediction based on \link{KNN.multioutput} function
#' \item{iter}: recursive prediction based on \link{KNN.multioutput} function
\item{mimo}: MIMO prediction based on \link{KNN.multioutput} function
\item{mimo.comb}: MIMO prediction based on \link{KNN.multioutput} function which combines a set of predictors based on different horizons and different starting points
\item{mimo.acf}: MIMO prediction based on \link{KNN.acf} function which combines a set of predictors based on different horizons and different starting points
\item{mimo.acf.lin}: MIMO prediction based on \link{KNN.acf.lin} function which combines a set of predictors based on different horizons and different starting points
\item{mimo.pls}: MIMO prediction based on \link{KNN.pls} function which combines a set of predictors based on different horizons and different starting points
}}
}
\value{
H step ahead predictions
}
\description{
multiplestepAhead
}
\details{
multiplestepAhead

Wrapper over a set of methods for multiple step ahead time series prediction
}
\examples{
## Multi-step ahead time series forecasting
library(pls)
t=seq(0,400,by=0.1)
N<-length(t)
H<-1500 ## horizon prediction
TS<-sin(t)+rnorm(N,sd=0.1)
TS.tr=TS[1:(N-H)]
N.tr<-length(TS.tr)
TS.ts<-TS[(N-H+1):N]
TS.tr=array(TS.tr,c(length(TS.tr),1))
E=MakeEmbedded(TS.tr,n=3,delay=0,hor=H,1)
X<-E$inp
Y<-E$out
N<-NROW(X)
ACF.lag<-5
Y.cont<-KNN.pls(X,Y,rev(TS.tr[(N.tr-H):N.tr]))
plot(t[(N-H+1):N],TS.ts)
lines(t[(N-H+1):N],Y.cont)
}
\author{
Gianluca Bontempi  \email{gbonte@ulb.ac.be}
}
\references{
\emph{Bontempi G. Ben Taieb S. Conditionally dependent strategies for multiple-step-ahead prediction in local learning, International Journal of Forecasting Volume 27, Issue 3, July–September 2011, Pages 689–699}
}

