% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotF.R
\name{plotF}
\alias{plotF}
\title{Plot the F density and cumulative distribution}
\usage{
plotF(N = 10, N2 = 20)
}
\description{
Plot F distribution
}
\references{
\url{mlg.ulb.ac.be}
}
\author{
Gianluca Bontempi  \email{gbonte@ulb.ac.be}
}
