% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featuresel.R
\name{rankrho}
\alias{rankrho}
\title{rankrho}
\usage{
rankrho(X, Y, nmax = 5, regr = FALSE)
}
\arguments{
\item{X:}{input dataset}

\item{Y:}{output dataset}

\item{nmax:}{number of top returned features}
}
\value{
Indices of \code{nmax} top ranked features
}
\description{
Ranking filter based on mutual information
}
\details{
ranking

Ranking filter based on mutaul information
}
\examples{
N<-100
n<-5
neff<-3
R<-regrDataset(N,n,neff,0.1,seed=0)
X<-R$X
Y<-R$Y
real.features<-R$feat
ranked.features<-rankrho(X,Y,nmax=3)
}
\references{
Handbook \emph{Statistical foundations of machine learning} available in \url{https://tinyurl.com/sfmlhf }
}
\author{
Gianluca Bontempi  \email{Gianluca.Bontempi@ulb.be}
}
